eclipse: Fix build issue with ADT Plugin v22.
Private libraries have to be explicitly exported now.  Also realized
that ACRA was being redundantly declared.  It's already in private
libraries, so that's all we need to add.  ADT v22 also wants to add
the dependencies path to this file as well... so we'll just go with it.
Doesn't seem to break the build for devs using earlier ADT versions.