front: Cleanup video plugin names, strings, credits for consistency.
Renamed our glide port "gles2glide64" to distinguish it from the
upstream version "glide64mk2".  Also follows the current directory name
for the plugin and mirrors the naming of the other video ports in the
app, i.e. prefix with gles2 to indicate an OpenGL ES 2.0 port of an
original library.

Removed the credits for the plugins. Would be better to just post these
on the forum sticky where there's more room and we can link to upstream
sources, etc.

Updated the plugin version names to match what I *think* these plugins
are derived from.