front/ae-bridge: Re-implement screen sizing.
TODO:
 - Use correct aspect ratio for PAL (do we even support PAL?)

This commit changes the way the video sizing is implemented. Previously
the core was given a rendering context that spanned the full extent of
screen space provided by Android's layout manager.  It was the duty of
the video plugin to define a viewport within that context to render to.
This required a significant amount of custom code to be added to the
video plugins just for the android edition.  There were also some arti-
facts with this method for Rice on some devices (i.e. junk graphics in
the margins).

Here, the alternative approach is to do the render context downsizing
on the Java side using the Android layout management APIs.  Instead of
passing a GameSurface that fills the screen, we make the GameSurface
only extend to the final desired rendering size.  The video plugins are
thus guaranteed not to draw in the pillarbox/letterbox margins, and
need not perform those sizing calculations.

The benefits are
 - No customization of the video plugins required
    - Simplifies integration of new video plugins
    - Easier to maintain
    - Easy to add new video features (resolution, hor. position, etc.)
 - Fixes "junk graphics in the margin" issues on Rice
 - Adds screen position feature to glide