Initialize audiobuffer before mixing it against our data
SDL doesn't guarantee that the target audio buffer passed in its callback is
initialized to zero (silence). SDL_MixAudio may now try to mix the random data
from the dst buffer and the new data from the src buffer together and creating
distorted sound in this process.

Initializing the destination buffer to zero avoids this problem.