front: Test vibrate permission directly rather than testing OUYA.
In case OUYA version is built with vibrate permission, or in case the
permission is removed from the manifest for a non-OUYA reason.

Note that we are setting vibrator objects to null if vibrate permission
is not granted, and then testing only for null before calling methods
that require the permission.  Leads to clean code but might be riskier
for code maintenance...