front: Changed "OUYA mode" to "Big-screen mode".
Like Steam's "Big Picture" mode.

Big-screen mode is useful for anyone running the app on a television or
monitor, not just OUYA users.  Also makes the code a little more clear,
avoiding confusion with OUYAInterface.IS_OUYA_HARDWARE, which tests
exclusively for the presence of the OUYA system.

Also consolidated the ActionBar check into a UserPrefs field for better
readability.