Allow to replace patented S3TC algorithm with compatible txc_dxtn
The S3TC is patented in some countries which makes it hard for distribution.
Mesa is using an external library which is using a compatible algorithm called
S2TC. But also S3TC implementations of this library are available.

Distributions can use this option to be on the safe side regarding the provided
functionality.