gles2rice: Remove JNI call for RGBA_8888 setting.
We are already handling this setting on the Java side in the
GameSurface.createGLContext() and CoreInterfaceNative.createGLContext()
functions.  Doing in via JNI is redundant.

At this point, the only reason the RGBA_8888 JNI function still exists
is because gles2n64 still uses it.  That plugin is bypassing the Mupen
core API to configure the video (thus bypassing the Java context
creation functions), instead calling SDL directly for context creation.