front: Added intent filter (ROMs can be launched from file managers).
A quick implementation that merely sets the selected ROM in the user
preferences.

In future versions of the app, it might make more sense to forward the
ROM info through the chain of intent data, rather than through the user
preferences.  But this implementation is fine for the current version.