SDL: Remove all trace of SDL 1.3.
SDL 1.3 is no longer supported, and has been superseded by SDL 2.0.
We've been carrying this code around for awhile, in case we needed to
do any regression testing.  But I think now is as good time as any to
cut the ties, to minimize hassle as we push towards app version 3.0.