front: Added 'crop' setting for video scaling.
Addresses Issue #197

This setting is useful for games like Jet Force Gemini that have in-
game widescreen settings.  The ROM itself letterboxes the video (adds
black bars to the top and bottom).  Adding a 'crop' option to the
scaling mode allows us to stretch to the full width of the screen and
crop the unneeded black bars that the ROM adds.