front: Simplify header CRC handling.
The complicated CRC check can be replaced with a simple String.format
operation. Note that the new implementation isn't identical to the old
in every way: before you could get CRC's like "0 0" from the native
side, which would then be turned into a null after the check.  Now, if
given the same file, the final CRC would be "00000000 00000000" instead
of null.  However, at the end of the day, this makes no difference to
the downstream code; it will just fail to find a cheat entry in either
case.