front: Remove unnecessary Java loading of native libraries.
The core loads and unloads the plugins automatically from the native
side, so there's no reason to load them here.  Loading from Java is
actually contributing to the lifecycle issues, because it prevents the
core from actually fully unloading them when the emulator shuts down.

For now, we still load the android-input plugin from Java, since we
initialize it using Java methods.  We can change this later if we want,
but for now it doesn't cause any problems.

A goal in future commits will be to eliminate the Java loading of the
front-end, core, and SDL libraries as well, so that their static vari-
able initializers will be invoked when relaunching the emulator.

Also removed the CoreInterfaceNative library loading wrappers.  If
exceptions are thrown, we have bigger problems to deal with and we
shouldn't just brush them under the rug.