front: Make game lifecycle handling more robust.
The emulator should only be started when the activity is resumed, the
surface exists, and the window has focus.  These states aren't
guaranteed to change in any particular order, so we simply track each
one and try to change the emulator state whenever one changes.  To
keep this stuff clean, we modularize it into a separate class.

The new implementation makes a few lines in CoreInterface superfluous,
so we remove them.

Note: There's still some weirdness when you hit the home button or
receive a call, and then resume the app.  The next commit will change a
whole lot of stuff related to the issue, so there's no point fixing it
yet.