front/ae-bridge: Remove system.exit calls, use normal activity cycle.
NOTE: On some versions of Android, if you quit the game activity (to
the play menu) and then resume/restart the game activity, you will get
a segfault (hard crash) as the game activity starts up.

I am intentionally committing this to demonstrate the need for native-
side library loading and unloading.  If you load a native library
through Java's System.loadLibrary function, there is no way to unload
it later.  This causes problems if the native code contains static
variables with static initializers (which is the case for us).  On
subsequent launches of the activity, the old library instance is re-
used rather than reloaded, and the static variables are thus not re-
initialized.  Thus you cannot reliably restart the activity in a fresh
state.

To get around the problem, we must load and unload the libraries on the
native side using dlopen, dlsym, and dlclose.  For a great discussion,
see http://stackoverflow.com/q/10649119/254218

Upcoming commits will address this problem, but I wanted to commit this
version as a frame of reference for subsequent discussion.