ae-bridge: Remove obsolete do_Start hack.
This flag no longer does anything.  Because the ae-imports library is
dynamically loaded and unloaded, do_Start will always have a value of 1
when it is first tested in the while loop in the front-end main func.

Also, we removed the reset function from the mapping screen way back at
commit 0b5083a5c3effa5a82762c3d778b926339dbf543, so no one has been
using this function anyhow.