ae-bridge: Remove onResize callback, move to native code.
For us, this callback has no effect on screen sizing, because it's
coordinated elsewhere between the core and SDL, and is specified dir-
ectly in the mupen config file.  The only reason we need this callback
is to initialize SDL's internal pixelformat state.  It's much cleaner
to just do this directly from the native side during the startup
sequence.  No need to involve the Java side or the activity/surface
lifecycle stuff.

Note that before this commit, the RGB565 SDL pixel format was always
being called, regardless of the user RGBA8888 setting.  I'm not sure if
the value matters much here; it just can't be the initial value
SDL_PIXELFORMAT_UNKNOWN.