ae-bridge: Internalize lib loading and delay unloading.
No need for the Java client to load and unload the libraries.  We can
keep this all internal to the ae-bridge.  Reduces the number of JNI
functions and minimizes complexity.

Also, we don't unload the libraries after the core thread exits.
Instead, we just unload stale library instances during the startup
sequence.  This avoids any potential race conditions between unloading
and the core commands during the shutdown sequence.