front: Fix "singleTask" activity behavior.
This mainly addresses this scenario:
 - User playing game or navigating the main menus
 - User presses home or gets interrupted by another app/phone call
 - User returns to home screen and launches a file manager app
 - User navigates to a different ROM and launches it from the manager

The desired behavior would be for the old GameActivity and PlayMenuAc-
tivity to be destroyed, and return the user to the main menu with the
new game selected.  This wasn't happening, because the "singleTask"
attribute was attached to the wrong xml element.  Must be attached to
an activity rather than the application.