ae-bridge: Add screenshot command to JNI.
I wasn't actually able to capture any screenshots with this (just black
images)... but I'll commit it anyhow in case another dev wants to make
it work.  Would definitely be useful, so that we can attach screenshots
to state-saves in version 3.  Also could make bug reporting a little
easier for the masses.

If you do want to try to get this working, be sure to specify a valid
target directory in CoreInterface.syncConfigFiles().