bugfix: Fix initial game menu state for some Android versions.
On my gingerbread device, the state callback was occurring before the
options menu was created, so the UI wasn't initialized with the proper
slot and speed.  Might be a version thing, or more likely just a race
condition.  In any case, this commit ensures that the UI initializes to
the correct info regardless of event order.

Also fixed the fallback speed value to 100 when ae-exports hasn't been
initialized yet.