xperia-touchpad: Fix hang on destroy.
Now that the lifecycle is fixed, we're finally exercising the shutdown
machinery in the xperia-touchpad library.  It seems the wrong pthread
function was being called.  app_thread_entry needs to raise the
destroyed flag and *wait* for OnDestroy to complete before it can con-
tinue and destroy the mutex and other objects.