xperia-touchpad: Re-implement using android_native_app_glue.
Major simplification of the native xperia-touchpad implementation.

Not just a cosmetic/maintenance fix. This also fixes a thread orphaning
bug.  Before, every time the user quit to the main menu from the game,
one of the threads was not destroyed properly and was left dangling.
So every time the user started and finished the game activity, another
dead thread would be added to the process. We never noticed the problem
before, since System.exit() force-killed these orphaned threads.  But
with the new lifecycle fixes, this bug became evident.

One note: If you build with NDK r9, the xperia touchpad and touchscreen
will spam the logcat with spurious errors.  They're not really errors,
but a new bug in the NDK.  If you build with NDK r8, you won't get the
spam.  See here for details:
http://stackoverflow.com/q/18316046/254218