r4300: Fix Perfect Dark crashes/freezes.
Had a look at this issue -
http://code.google.com/p/mupen64plus/issues/detail?id=553

Implemented the proposed fix on my own and that got rid of the spurious
crashes for me.  BTW - Misschief Makers is another game that reports
this warning "two events of same type interrupt queue". Anyway,
returning when that happens gets rid of the crashes in PD.

[cherry pick from
https://github.com/libretro/mupen64plus-libretro/commit/12133422f75aa3f88ef5decc820402e894371adc]

littleguy note: You might need to delete any pre-existing .eep and .mpk
files from the SlotSaves directory and reset the game the first time
you run with this fix.  More discussion here:
http://www.paulscode.com/forum/index.php?topic=1042