bugfix: Fix filepath regression (e.g. rice hi-res textures).
Bug introduced in commit
48e94450531fe6af307737cb3606b492e188d569

which broke two methods in the core API:

   - ConfigGetUserDataPath
   - ConfigGetUserCachePath

These paths can only be resolved through the Android customizations
that we've been using all along.  It wasn't obvious, however, because
the only code that calls these methods are:
   - Rice uses ConfigGetUserDataPath for hi-res textures & tex dumps
   - Glide uses both, but only if TEXTURE_FILTER is defined (which it
     currently isn't)

Note that ConfigGetUserConfigPath and ConfigGetSharedDataPath never
broke because they are obtained by the core via command-line arg and
config file, respectively.