front: Hide system buttons again when window focus returns.
Fixes a minor UI nit.  Before, if the window focus was lost (e.g. call
received, pressed home button, etc.), then when the user returned to
mupen, the navigation buttons at the bottom of the screen were no
longer in "lights out" mode.  This just ensures that the lights-out
mode is restored when focus returns to the app.

This fix becomes more important once immersive mode is implemented,
because then it is much more noticeable to the user when mupen regains
focus.