ae-bridge: Fix regression where launch crashed after settings changed.
This commit effectively reverts a regression tracked down to commit
2e4f3c70c477a1ac96745d061f141af2066d8a97

Somehow the order of loading (input-android vs other libraries) and/or
the thread from which they are loaded makes a difference.

The bug can be reproduced by launching emulator, quitting, changing a
user setting (e.g. audio plugin), then relaunching the emulator.  The
app would crash hard due to a segfault.  The crash only occurred when
a user setting was changed; never occurred if the settings didn't
change.