front: Always use framelimiter when not using dummy audio.
Enabling framelimiter (the upstream default) minimizes latency with the
audio-sdl plugin. Disabling the framelimiter is only useful for bench-
marking framerate with dummy audio.

In my testing with older devices, I see no benefit to disabling frame-
limiter as a means of increasing framerate.  My guess is that this
commit won't negatively affect older devices, though it would be nice
to know for sure.