front: Remove zipped-rom feature for now.
We'll probably do it differently once the carousel is finished. e.g.
user touches 'refresh' and then the app scans for .z64, .v64, .n64, and
(optionally) .zip files. If it finds zips with roms in them, it unpacks
them once to a stable (non-tempor​ary) location.  No repeated unpacking
all the time.