front: Replace controller mapping with controller profiles system.
This is necessarily a big commit because it's a paradigm shift in the
way controllers are configured.  Instead of visiting the mapping screen
every time the user wants to change controller setup, they simply
change the controller profile.  There are a large number of built-in
profiles for ease of use, and the user can still define custom mappings
however they want.  It's easier at a glance to see how each controller
is configured, and fewer clicks through the configuration.

This also eliminates the controller toggle switch in the input screen,
which was not good for big-screen (controller-only) navigation.

This is also much more amenable to per-rom configuration.  The per-rom
override will only need to override a single profile name, not a whole
assortment of button mappings, deadzone, sensitivity, etc.