tools: Add a whitespace only pull script
It seems that some whitespace differences are popping up from time to time and
get never forwarded to upstream. This makes it unnecessary hard to merge
changes from upstream back to mupen64plus. The tool pull-upstream-whitespace.sh
can be used to get only the whitespace differences from upstream and not the
actual changes/commits.

This script can fail and the caller has to find the "*.rej" files and manually
merge the differences. It is also important to check for unadded/not removed
files. Also a diff should be made before committing in such a state:

git diff --ignore-all-space HEAD