front: Remove workaround for defective Nyko PlayPad firmware.
This hack fix is starting to interfere with too many other controller
models that actually implement HID properly.  M.O.J.O is the latest
victim of Nyko's blunder.

The UI is mature enough now that removing the hack shouldn't cause too
much pain.  The Nyko PlayPad left stick is mapped to N64 analog stick
by default, and the default can be recovered easily in the mapping
screen's menus.  The only pain this could cause is if a Nyko owner
wants to re-map the left stick.  In really desperate circumstances, the
profile can be edited manually with a text editor.  I can live with
that for now.