front: Fix bug where touchpad could be unintentionally enabled.
Bug introduced in commit
17539be993ccbd44d3ff60b0e3423a3b287e96f6
and exacerbated when global prefs menu added to PlayMenuActivity.

Bug occurs when user resets global preferences.  If they launch a game
before restarting the app, the touchpad would always be enabled, since
the hardware test to disable it for non-Xperia-PLAY devices only hap-
pens in SplashActivity.

The bug breaks analog controls when using a peripheral controller,
because GameActivityXperiaPlay ignores joystick events. (It only
listens to touch and button events.)  There might be a performance
difference between the GameActivity* variants as well.