front: Do not auto-set some preferences for big-screen mode.
No need to be heavy handed.  Better safe than sorry.  The visibility of
touchscreen-related elements will probably be controlled differently
anyhow in version 3... so just stay out of the way for now.

As a general rule, don't rely on SplashActivity to initialize anything
that is user-settable.  Use it only to initialize non-user-configurable
stuff like static databases.