front: Add base class for PreferenceActivity-like profile editors.
Add an abstract base class for implementing simple list-based profile
editor activities. This class is meant to be subclassed the same way
that a PreferenceActivity would be, and the process for defining the
structure and behavior is virtually identical. That is, the hierarch-
ical menu structure and default values are defined using an xml re-
source, using identical syntax to that used for a PreferenceActivity.
Similarly, the behavior of the activity (lifecycle, inflation, menu
visibility, change notifications, etc.) is implemented just as it would
be for a PreferenceActivity. It is thus trivial to migrate implementa-
tion between a PreferenceActivity (backed by a SharedPreferences
object) and a ProfileActivity (backed by a ConfigSection object).

The one important caveat is that this class persists all settings as
Strings. This means that some of the standard Preference classes must
be extended a bit. For example, the CheckBoxPreference class persists
to boolean. Therefore, to make a checkbox-style preference that works
with ProfileActivity, extend CheckBoxPreference and override the
persistBoolean and getPersistedBoolean methods.