front: Revert activity transitions back to defaults.
We used to restart PreferenceActivities to refresh their menus in some
cases.  At that time, it was preferable to disable animations so that
the effect wasn't so jarring and so that several activities looked like
a single one when traversing menus.  But now, we no longer restart
activities for this purpose, and the transition animations actually
make sense now, to match standard Android patterns.