tools: Temporarily remove pull-upstream scripts.
The tools removed here will be coming back soon in a slightly altered
form.  I'm removing them now for safety as we administer to the git
repos.

We are migrating over to new upstream repositories.  We have created
official forks of the upstream repositories, and will be pushing our
android customizations to them from now on.  That will make it very
easy for other devs to see the changes we made to the upstream plugins
for the android edition.  Also makes it easy for other embedded system
ports (libretro, pandora) to pull just the plugins of interest out of
mupen64plus-ae.