front: Restructure CheatBlock/CheatLine hierarchy.
Before, CheatLine had two personalities.  It either represented a line
of text in the CheatFile, or it wrapped a CheatBlock object.  A flag was
used to discern the personality of an instance.

Now, CheatLine is a sibling to CheatBlock, and both descend from an
abstract base class, CheatElement.  CheatLines now only represent CRC
lines, goodname lines, and comment lines.