front: Simplify cheat file parsing, improve readability.
No performance penalty observed. If anything, there seems to be a very
slight boost.

Note that the field strings are no longer trimmed in this new implement-
ation.  However, that shouldn't be an issue since the whitespace would
have been in the original file too (and never gave us problems there).