front: Tweak cheat options regex a bit.
Eliminate "catastrophic backtracking" using possessive quantifiers (+).
Also, this version is a bit more efficient in that it takes fewer steps
to find a match, though the difference is hardly noticeable in practice.

Catastrophic backtracking can occur with malformed cheat lines, such as
 - omitting the final quotation mark on a cheat line containing options
 - adding a backslash before the final quotation mark on such a line

In practice, this adds a few hundred milliseconds to the parse time for
each malformed line.  Not much, but still good to avoid it.