front: Name game-specific data directory using immutable header info.
Header name is not always as user friendly as GoodName, but the header
name is immutable for a given MD5, which is probably more important in
the long run for not borking user data.

Using the header also eliminates the dependency on the RomDetails class,
which is sometimes costly to use.