front: Change default vertical position from center to top.
This setting generally only matters when playing the game in portrait
orientation.  Making the default be "top" gives plenty of space for the
touchscreen controls in the lower portion of the screen, yielding a lay-
out like the original Gameboy.  My guess is that most people who use the
game in portrait orientation would prefer it this way.  I certainly do.