tools: Update pull-upstream script.
We are now employing a git-submodule-like policy for synchronizing the
upstream components.  That is, all changes to the upstream modules in
the mupen64plus-ae repository must be made on the upstream side first,
then pulled downstream using this script.  Any manual modifications made
from the downstream side will be overwritten the next time this script
is run.  That is very much intentional, as it enforces the top-down
philosophy we are now adopting.

Developers may still manually modify the upstream modules from the down-
stream side, as long as they do it in a branch that never gets merged
back to master.  That is also very much intentional; it should still be
easy for devs to experiment with upstream code while working in the
downstream project.

Note that we are not actually using git-submodule, due to its infamous
fragility.  We do not want to require our contributors to be git super-
users.

The approach used here is much more robust than git-submodule, since the
upstream code is always retained in the downstream repository's history.
The approach used here simply emulates what a git superuser would do
using only basic git and shell operations.  That means that any mistakes
created by this commit can be fixed using basic git and shell commands.
No black magic or human sacrifices are needed in a worst-case scenario.