Drag and drop editing for the touchscreen controls
Things that are changed with this commit:

- added drag and drop editing for the on-screen controller layout.
Right now the values are clamped to the nearest 5% to match the
behavior of the edit dialog, although I feel like we should allow any %
for that.
- tapping on a button plays the standard button sound effect, as I felt
that was missing.
- made it possible to swipe from the top or bottom of the screen to
show the back/home/menu buttons or status bar without *also* triggering
the edit dialog for a button that was under your finger, since that
dialog prevented the user from being able to use the menus at the top
of the screen.