Ignore touches along all four edges
In immersive mode on KitKat there’s no way to ignore edge swipes, which
means swiping to reveal the UI will also move any buttons under your
finger. This is undesirable, so add a small dead zone for touches along
the edges. Fortunately buttons are much larger than 10 points so this
should not interfere with the editor too much.