Reimplemented grid item spacing
Cleaned up the code a bit and changed it to get the values for the
gallery thumbnail sizes from the dimens.xml file.

GridLayoutManager doesn’t seem to have built-in support for item
spacing, so I had to reimplement it manually by adding spacing around
the items and the grid view. Also cleaned up the math for the column
calculation.

Disable clip to padding on the grid layout