front: Replace UserPrefs getter with final field.
Using final fields is preferred because they are all loaded at the same
time OnCreate(), helping to ensure the whole UI is working with the same
user settings.  Dynamic getters should only be used when there's a
dynamic setter, which in turn should only where the built-in sharedprefs
mechanisms aren't suitable.