Fix auto-build script
The dependency configuration step was generating untracked files, which prevented subsequent calls to git checkout.  Added a line to clean up untracked files after each build.  Also, moved dependency configuration inside the block where the builds happen (so that it only executes in the same conditions that the cleanup executes).