front: Log the version of the GLES context that was created.
According to Android documentation, the device will always create the
highest possible GLES context that it supports.  So even if SDL requests
a 2.0 context, an NVIDIA Shield (for example) will produce a 3.1
context.  Thus we don't need to explicitly request a 3.1 context.

We log the created version here, as proof that the highest supported
context was created.