gles3: Copy headers and libraries from NDK-r10d.
(message amended by littleguy77)

To get GLES 3.1 support, we need to bump APP_PLATFORM to android-21 in
Application.mk.  Unfortunately, pre-Lollipop devices are not binary-
compatible with native code built against android-21, due to changes in
the platform sources.  The emulator will crash as soon as it is launched
on an older device.  See http://stackoverflow.com/a/27093163/254218 for
example.

We could always publish multiple APKs but this adds more complexity to
the build process.  More importantly, we'd have to distribute several
APKs every time we wanted to push test builds to the forum or devs
wanted to share works in progress.  The build complexity also adds
another barrier to new contributors.

In this case, the GLES headers are only a few files and rarely change.
They can always be diffed against the current NDK if there is ever a
concern about corruption.  Same for the binary .so files.  So instead of
leaning on the built-in binaries in android-21, we copy them into our
project so that we can stay at android-9.