Front: Don't stop emulation if surfaceDestroyed If surfaceDestroyed: - pause emulation (if not already paused) - notify SDL that the surface has been destroyed by sending the pause signal - unbind the previously-created EGL context and destroy the window surface
If surfaceChanged (Resume emulation):
- notify SDL that the surface is ready by sending the resume signal
- upon resume SDL will call createGLContext(Java) and re-create the EGL context (see Android_PumpEvents in SDL_androidevents.c)
- resume emulation when EGL context is ready

If surfaceChanged (Startup emulation):
- notify SDL that the surface is ready by sending the resume signal (useless)
- On SDL_SetVideoMode SDL will call createGLContext(Java) and create the EGL context
- emulation starts

If onDestroy stop emulation.

Autosave are now only performed upon exit.  The main reason for this is that savestates sometimes
failed to be executed when pausing emulation and is subsequently only performed when resuming.

SDL_ANDROID_BLOCK_ON_PAUSE must be undefined to allow the core to call SDL_PumpEvents cyclically